support MSI/MSI-X (not kernel option)

+++ sys/arch/x86/include/pci_machdep.h
@@ -37,7 +37,7 @@
  * Types provided to machine-independent PCI code
  * See also i82093var.h to find out pci_intr_handle_t's bitfield.
  */
-typedef int pci_intr_handle_t;
+typedef uint64_t pci_intr_handle_t;
 
 #include <x86/pci_machdep_common.h>
